russh::client

Trait Handler

pub trait Handler: Sized + Send {
    type Error: From<Error> + Send + Debug;

    // Provided methods
    fn auth_banner(
        &mut self,
        banner: &str,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn check_server_key(
        &mut self,
        server_public_key: &PublicKey,
    ) -> impl Future<Output = Result<bool, Self::Error>> + Send { ... }
    fn kex_done(
        &mut self,
        shared_secret: Option<&[u8]>,
        names: &Names,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn channel_open_confirmation(
        &mut self,
        id: ChannelId,
        max_packet_size: u32,
        window_size: u32,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn channel_success(
        &mut self,
        channel: ChannelId,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn channel_failure(
        &mut self,
        channel: ChannelId,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn channel_close(
        &mut self,
        channel: ChannelId,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn channel_eof(
        &mut self,
        channel: ChannelId,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn channel_open_failure(
        &mut self,
        channel: ChannelId,
        reason: ChannelOpenFailure,
        description: &str,
        language: &str,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_forwarded_tcpip(
        &mut self,
        channel: Channel<Msg>,
        connected_address: &str,
        connected_port: u32,
        originator_address: &str,
        originator_port: u32,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_forwarded_streamlocal(
        &mut self,
        channel: Channel<Msg>,
        socket_path: &str,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_agent_forward(
        &mut self,
        channel: Channel<Msg>,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn should_accept_unknown_server_channel(
        &mut self,
        id: ChannelId,
        channel_type: &str,
    ) -> impl Future<Output = bool> + Send { ... }
    fn server_channel_open_unknown(
        &mut self,
        channel: Channel<Msg>,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_session(
        &mut self,
        channel: Channel<Msg>,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_direct_tcpip(
        &mut self,
        channel: Channel<Msg>,
        host_to_connect: &str,
        port_to_connect: u32,
        originator_address: &str,
        originator_port: u32,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_direct_streamlocal(
        &mut self,
        channel: Channel<Msg>,
        socket_path: &str,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn server_channel_open_x11(
        &mut self,
        channel: Channel<Msg>,
        originator_address: &str,
        originator_port: u32,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn data(
        &mut self,
        channel: ChannelId,
        data: &[u8],
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn extended_data(
        &mut self,
        channel: ChannelId,
        ext: u32,
        data: &[u8],
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn xon_xoff(
        &mut self,
        channel: ChannelId,
        client_can_do: bool,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn exit_status(
        &mut self,
        channel: ChannelId,
        exit_status: u32,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn exit_signal(
        &mut self,
        channel: ChannelId,
        signal_name: Sig,
        core_dumped: bool,
        error_message: &str,
        lang_tag: &str,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn window_adjusted(
        &mut self,
        channel: ChannelId,
        new_size: u32,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn adjust_window(&mut self, channel: ChannelId, window: u32) -> u32 { ... }
    fn openssh_ext_host_keys_announced(
        &mut self,
        keys: Vec<PublicKey>,
        session: &mut Session,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn disconnected(
        &mut self,
        reason: DisconnectReason<Self::Error>,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }

}

A client handler. Note that messages can be received from the server at any time during a session.

You must at the very least implement the check_server_key fn. The default implementation rejects all keys.

Note: this is an async trait. The trait functions return impl Future, and you can simply define them as async fn instead.

Required Associated Types

type Error: From<Error> + Send + Debug

Provided Methods

fn auth_banner( &mut self, banner: &str, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server sends us an authentication banner. This is usually meant to be shown to the user, see RFC4252 for more details.

fn check_server_key( &mut self, server_public_key: &PublicKey, ) -> impl Future<Output = Result<bool, Self::Error>> + Send

Called to check the server’s public key. This is a very important step to help prevent man-in-the-middle attacks. The default implementation rejects all keys.

fn kex_done( &mut self, shared_secret: Option<&[u8]>, names: &Names, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when key exchange has completed.

This callback provides access to the raw shared secret from the KEX, which is useful for protocols that derive additional keys from the SSH shared secret (e.g., for secondary encrypted channels).

The names parameter contains all negotiated algorithms (kex, cipher, mac, etc.).

Security Warning: The shared secret is sensitive cryptographic material. Handle it with care and zero it after use if stored.

Arguments

kex_algorithm - Name of the key exchange algorithm used

shared_secret - The raw shared secret bytes from the key exchange. For some algorithms (like none), this may be None.

names - The negotiated algorithm names

session - The current session

fn channel_open_confirmation( &mut self, id: ChannelId, max_packet_size: u32, window_size: u32, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server confirmed our request to open a channel. A channel can only be written to after receiving this message (this library panics otherwise).

fn channel_success( &mut self, channel: ChannelId, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server signals success.

fn channel_failure( &mut self, channel: ChannelId, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server signals failure.

fn channel_close( &mut self, channel: ChannelId, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server closes a channel.

fn channel_eof( &mut self, channel: ChannelId, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server sends EOF to a channel.

fn channel_open_failure( &mut self, channel: ChannelId, reason: ChannelOpenFailure, description: &str, language: &str, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server rejected our request to open a channel.

fn server_channel_open_forwarded_tcpip( &mut self, channel: Channel<Msg>, connected_address: &str, connected_port: u32, originator_address: &str, originator_port: u32, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens a channel for a new remote port forwarding connection

fn server_channel_open_forwarded_streamlocal( &mut self, channel: Channel<Msg>, socket_path: &str, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

fn server_channel_open_agent_forward( &mut self, channel: Channel<Msg>, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens an agent forwarding channel

fn should_accept_unknown_server_channel( &mut self, id: ChannelId, channel_type: &str, ) -> impl Future<Output = bool> + Send

Called when the server attempts to open a channel of unknown type. It may return true, if the channel of unknown type should be accepted. In this case, Handler::server_channel_open_unknown will be called soon after. If it returns false, the channel will not be created and a rejection message will be sent to the server.

fn server_channel_open_unknown( &mut self, channel: Channel<Msg>, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens an unknown channel.

fn server_channel_open_session( &mut self, channel: Channel<Msg>, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens a session channel.

fn server_channel_open_direct_tcpip( &mut self, channel: Channel<Msg>, host_to_connect: &str, port_to_connect: u32, originator_address: &str, originator_port: u32, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens a direct tcp/ip channel (non-standard).

fn server_channel_open_direct_streamlocal( &mut self, channel: Channel<Msg>, socket_path: &str, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens a direct-streamlocal channel (non-standard).

fn server_channel_open_x11( &mut self, channel: Channel<Msg>, originator_address: &str, originator_port: u32, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server opens an X11 channel.

fn data( &mut self, channel: ChannelId, data: &[u8], session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server sends us data. The extended_code parameter is a stream identifier, None is usually the standard output, and Some(1) is the standard error. See RFC4254.

fn extended_data( &mut self, channel: ChannelId, ext: u32, data: &[u8], session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server sends us data. The extended_code parameter is a stream identifier, None is usually the standard output, and Some(1) is the standard error. See RFC4254.

fn xon_xoff( &mut self, channel: ChannelId, client_can_do: bool, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

The server informs this client of whether the client may perform control-S/control-Q flow control. See RFC4254.

fn exit_status( &mut self, channel: ChannelId, exit_status: u32, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

The remote process has exited, with the given exit status.

fn exit_signal( &mut self, channel: ChannelId, signal_name: Sig, core_dumped: bool, error_message: &str, lang_tag: &str, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

The remote process exited upon receiving a signal.

fn window_adjusted( &mut self, channel: ChannelId, new_size: u32, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the network window is adjusted, meaning that we can send more bytes. This is useful if this client wants to send huge amounts of data, for instance if we have called Session::data before, and it returned less than the full amount of data.

fn adjust_window(&mut self, channel: ChannelId, window: u32) -> u32

Called when this client adjusts the network window. Return the next target window and maximum packet size.

fn openssh_ext_host_keys_announced( &mut self, keys: Vec<PublicKey>, session: &mut Session, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server signals success.

fn disconnected( &mut self, reason: DisconnectReason<Self::Error>, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Called when the server sent a disconnect message

If reason is an Error, this function should re-return the error so the join can also evaluate it

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.