russh

Module keys

This crate contains methods to deal with SSH keys, as defined in crate Russh. This includes in particular various functions for opening key files, deciphering encrypted keys, and dealing with agents.

The following example shows how to do all these in a single example: start and SSH agent server, connect to it with a client, decipher an encrypted ED25519 private key (the password is b"blabla"), send it to the agent, and ask the agent to sign a piece of data (b"Please sign this", below).

 use russh::keys::*;
 use futures::Future;

 #[derive(Clone)]
 struct X{}
 impl agent::server::Agent for X {
     fn confirm(self, _: std::sync::Arc<PrivateKey>) -> Box<dyn Future<Output = (Self, bool)> + Send + Unpin> {
         Box::new(futures::future::ready((self, true)))
     }
 }

 const PKCS8_ENCRYPTED: &'static str = "-----BEGIN ENCRYPTED PRIVATE KEY-----\nMIGjMF8GCSqGSIb3DQEFDTBSMDEGCSqGSIb3DQEFDDAkBBAWQiUHKoocuxfoZ/hF\nYTjkAgIIADAMBggqhkiG9w0CCQUAMB0GCWCGSAFlAwQBKgQQ83d1d5/S2wz475uC\nCUrE7QRAvdVpD5e3zKH/MZjilWrMOm6cyI1LKBCssLztPyvOALtroLAPlp7WYWfu\n9Sncmm7u14n2lia7r1r5I3VBsVuH0g==\n-----END ENCRYPTED PRIVATE KEY-----\n";

 #[cfg(unix)]
 fn main() {
    env_logger::try_init().unwrap_or(());
    let dir = tempfile::tempdir().unwrap();
    let agent_path = dir.path().join("agent");

    let mut core = tokio::runtime::Runtime::new().unwrap();
    let agent_path_ = agent_path.clone();
    // Starting a server
    core.spawn(async move {
        let mut listener = tokio::net::UnixListener::bind(&agent_path_)
            .unwrap();
        russh::keys::agent::server::serve(tokio_stream::wrappers::UnixListenerStream::new(listener), X {}).await
    });
    let key = decode_secret_key(PKCS8_ENCRYPTED, Some("blabla")).unwrap();
    let public = key.public_key().clone();
    core.block_on(async move {
        let stream = tokio::net::UnixStream::connect(&agent_path).await?;
        let mut client = agent::client::AgentClient::connect(stream);
        client.add_identity(&key, &[agent::Constraint::KeyLifetime { seconds: 60 }]).await?;
        client.request_identities().await?;
        let buf = b"signed message";
        let sig = client.sign_request(&public, None, russh_cryptovec::CryptoVec::from_slice(&buf[..])).await.unwrap();
        // Here, `sig` is encoded in a format usable internally by the SSH protocol.
        Ok::<(), Error>(())
    }).unwrap()
 }

 #[cfg(not(unix))]
 fn main() {}

Re-exports

pub use key::PrivateKeyWithHashAlg;

pub use known_hosts::check_known_hosts;

pub use known_hosts::check_known_hosts_path;

pub use signature;

pub use ssh_encoding;

pub use ssh_key;

Modules

agent

OpenSSH agent protocol implementation

key

known_hosts

openssh

pkcs5

pkcs8

Structs

Certificate

OpenSSH certificate as specified in PROTOCOL.certkeys.

PrivateKey

SSH private key.

PublicKey

SSH public key.

Enums

Algorithm

SSH key algorithms.

EcdsaCurve

Elliptic curves supported for use with ECDSA.

Encryption

AES encryption key.

Error

HashAlg

Hashing algorithms a.k.a. digest functions.

Traits

PublicKeyBase64

Functions

decode_openssh

Decode a secret key given in the OpenSSH format, deciphering it if needed using the supplied password.

decode_pkcs5

Decode a secret key in the PKCS#5 format, possibly deciphering it using the supplied password.

decode_secret_key

Decode a secret key, possibly deciphering it with the supplied password.

encode_pkcs8_pem

encode_pkcs8_pem_encrypted

load_openssh_certificate

Load a openssh certificate

load_public_key

Load a public key from a file. Ed25519, EC-DSA and RSA keys are supported.

load_secret_key

Load a secret key, deciphering it with the supplied password if necessary.

parse_public_key_base64

Reads a public key from the standard encoding. In some cases, the encoding is prefixed with a key type identifier and a space (such as ssh-ed25519 AAAAC3N...).