russh::keys::agent::client

Struct AgentClient

pub struct AgentClient<S: AgentStream> { /* private fields */ }

SSH agent client.

Implementations

impl<S: AgentStream + Send + Unpin + 'static> AgentClient<S>

pub fn dynamic( self, ) -> AgentClient<Box<dyn AgentStream + Send + Unpin + 'static>>

Wraps the internal stream in a Box, allowing different client implementations to have the same type

pub fn into_inner(self) -> Box<dyn AgentStream + Send + Unpin + 'static>

impl<S: AgentStream + Unpin> AgentClient<S>

pub fn connect(stream: S) -> Self

Build a future that connects to an SSH agent via the provided stream (on Unix, usually a Unix-domain socket).

impl AgentClient<UnixStream>

pub async fn connect_uds<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Connect to an SSH agent via the provided stream (on Unix, usually a Unix-domain socket).

pub async fn connect_env() -> Result<Self, Error>

Connect to an SSH agent specified by the SSH_AUTH_SOCK environment variable.

impl<S: AgentStream + Unpin> AgentClient<S>

pub async fn add_identity( &mut self, key: &PrivateKey, constraints: &[Constraint], ) -> Result<(), Error>

Send a key to the agent, with a (possibly empty) slice of constraints to apply when using the key to sign.

pub async fn add_smartcard_key( &mut self, id: &str, pin: &[u8], constraints: &[Constraint], ) -> Result<(), Error>

Add a smart card to the agent, with a (possibly empty) set of constraints to apply when signing.

pub async fn lock(&mut self, passphrase: &[u8]) -> Result<(), Error>

Lock the agent, making it refuse to sign until unlocked.

pub async fn unlock(&mut self, passphrase: &[u8]) -> Result<(), Error>

Unlock the agent, allowing it to sign again.

pub async fn request_identities(&mut self) -> Result<Vec<PublicKey>, Error>

Ask the agent for a list of the currently registered secret keys.

pub async fn sign_request( &mut self, public: &PublicKey, hash_alg: Option<HashAlg>, data: CryptoVec, ) -> Result<CryptoVec, Error>

Ask the agent to sign the supplied piece of data.

pub fn sign_request_base64( self, public: &PublicKey, hash_alg: Option<HashAlg>, data: &[u8], ) -> impl Future<Output = (Self, Result<String, Error>)>

Ask the agent to sign the supplied piece of data.

pub async fn sign_request_signature( &mut self, public: &PublicKey, hash_alg: Option<HashAlg>, data: &[u8], ) -> Result<Signature, Error>

Ask the agent to sign the supplied piece of data, and return a Signature.

pub async fn remove_identity(&mut self, public: &PublicKey) -> Result<(), Error>

Ask the agent to remove a key from its memory.

pub async fn remove_smartcard_key( &mut self, id: &str, pin: &[u8], ) -> Result<(), Error>

Ask the agent to remove a smartcard from its memory.

pub async fn remove_all_identities(&mut self) -> Result<(), Error>

Ask the agent to forget all known keys.

pub async fn extension(&mut self, typ: &[u8], ext: &[u8]) -> Result<(), Error>

Send a custom message to the agent.

pub async fn query_extension( &mut self, typ: &[u8], ext: CryptoVec, ) -> Result<bool, Error>

Ask the agent what extensions about supported extensions.