russh::server

Enum Msg

pub enum Msg {
    ChannelOpenAgent {
        channel_ref: ChannelRef,
    },
    ChannelOpenSession {
        channel_ref: ChannelRef,
    },
    ChannelOpenDirectTcpIp {
        host_to_connect: String,
        port_to_connect: u32,
        originator_address: String,
        originator_port: u32,
        channel_ref: ChannelRef,
    },
    ChannelOpenDirectStreamLocal {
        socket_path: String,
        channel_ref: ChannelRef,
    },
    ChannelOpenForwardedTcpIp {
        connected_address: String,
        connected_port: u32,
        originator_address: String,
        originator_port: u32,
        channel_ref: ChannelRef,
    },
    ChannelOpenForwardedStreamLocal {
        server_socket_path: String,
        channel_ref: ChannelRef,
    },
    ChannelOpenX11 {
        originator_address: String,
        originator_port: u32,
        channel_ref: ChannelRef,
    },
    TcpIpForward {
        reply_channel: Option<Sender<Option<u32>>>,
        address: String,
        port: u32,
    },
    CancelTcpIpForward {
        reply_channel: Option<Sender<bool>>,
        address: String,
        port: u32,
    },
    Disconnect {
        reason: Disconnect,
        description: String,
        language_tag: String,
    },
    Channel(ChannelId, ChannelMsg),
}

Variants

ChannelOpenAgent

Fields

channel_ref: ChannelRef

ChannelOpenSession

Fields

channel_ref: ChannelRef

ChannelOpenDirectTcpIp

Fields

host_to_connect: String

port_to_connect: u32

originator_address: String

originator_port: u32

channel_ref: ChannelRef

ChannelOpenDirectStreamLocal

Fields

socket_path: String

channel_ref: ChannelRef

ChannelOpenForwardedTcpIp

Fields

connected_address: String

connected_port: u32

originator_address: String

originator_port: u32

channel_ref: ChannelRef

ChannelOpenForwardedStreamLocal

Fields

server_socket_path: String

channel_ref: ChannelRef

ChannelOpenX11

Fields

originator_address: String

originator_port: u32

channel_ref: ChannelRef

TcpIpForward

Fields

reply_channel: Option<Sender<Option<u32>>>

Provide a channel for the reply result to request a reply from the server

address: String

port: u32

CancelTcpIpForward

Fields

reply_channel: Option<Sender<bool>>

Provide a channel for the reply result to request a reply from the server

address: String

port: u32

Disconnect

Fields

reason: Disconnect

description: String

language_tag: String

Channel(ChannelId, ChannelMsg)