russh::server

Struct Config

pub struct Config {

    pub server_id: SshId,
    pub methods: MethodSet,
    pub auth_rejection_time: Duration,
    pub auth_rejection_time_initial: Option<Duration>,
    pub keys: Vec<PrivateKey>,
    pub limits: Limits,
    pub window_size: u32,
    pub maximum_packet_size: u32,
    pub channel_buffer_size: usize,
    pub event_buffer_size: usize,
    pub preferred: Preferred,
    pub max_auth_attempts: usize,
    pub inactivity_timeout: Option<Duration>,
    pub keepalive_interval: Option<Duration>,
    pub keepalive_max: usize,
    pub nodelay: bool,

}

Configuration of a server.

Fields
server_id: SshId

The server ID string sent at the beginning of the protocol.

methods: MethodSet

Authentication methods proposed to the client.

auth_rejection_time: Duration

Authentication rejections must happen in constant time for security reasons. Russh does not handle this by default.

auth_rejection_time_initial: Option<Duration>

Authentication rejection time override for the initial “none” auth attempt. OpenSSH clients will send an initial “none” auth to probe for authentication methods.

keys: Vec<PrivateKey>

The server’s keys. The first key pair in the client’s preference order will be chosen.

limits: Limits

The bytes and time limits before key re-exchange.

window_size: u32

The initial size of a channel (used for flow control).

maximum_packet_size: u32

The maximal size of a single packet.

channel_buffer_size: usize

Buffer size for each channel (a number of unprocessed messages to store before propagating backpressure to the TCP stream)

event_buffer_size: usize

Internal event buffer size

preferred: Preferred

Lists of preferred algorithms.

max_auth_attempts: usize

Maximal number of allowed authentication attempts.

inactivity_timeout: Option<Duration>

Time after which the connection is garbage-collected.

keepalive_interval: Option<Duration>

If nothing is received from the client for this amount of time, send a keepalive message.

keepalive_max: usize

If this many keepalives have been sent without reply, close the connection.

nodelay: bool

If active, invoke set_nodelay(true) on client sockets; disabled by default (i.e. Nagle’s algorithm is active).