russh::server
Struct Handle

pub struct Handle { /* private fields */ }

Handle to a session, used to send messages to a client outside of the request/response cycle.

Implementations

impl Handle

pub async fn data( &self, id: ChannelId, data: CryptoVec, ) -> Result<(), CryptoVec>

Send data to the session referenced by this handler.

pub async fn extended_data( &self, id: ChannelId, ext: u32, data: CryptoVec, ) -> Result<(), CryptoVec>

Send data to the session referenced by this handler.

pub async fn eof(&self, id: ChannelId) -> Result<(), ()>

Send EOF to the session referenced by this handler.

pub async fn channel_success(&self, id: ChannelId) -> Result<(), ()>

Send success to the session referenced by this handler.

pub async fn channel_failure(&self, id: ChannelId) -> Result<(), ()>

Send failure to the session referenced by this handler.

pub async fn close(&self, id: ChannelId) -> Result<(), ()>

Close a channel.

pub async fn xon_xoff_request( &self, id: ChannelId, client_can_do: bool, ) -> Result<(), ()>

Inform the client of whether they may perform control-S/control-Q flow control. See RFC4254.

pub async fn exit_status_request( &self, id: ChannelId, exit_status: u32, ) -> Result<(), ()>

Send the exit status of a program.

pub async fn forward_tcpip(&self, address: String, port: u32) -> Result<u32, ()>

Notifies the client that it can open TCP/IP forwarding channels for a port.

pub async fn cancel_forward_tcpip( &self, address: String, port: u32, ) -> Result<(), ()>

Notifies the client that it can no longer open TCP/IP forwarding channel for a port.

pub async fn channel_open_session(&self) -> Result<Channel<Msg>, Error>

Request a session channel (the most basic type of channel). This function returns Ok(..) immediately if the connection is authenticated, but the channel only becomes usable when itโs confirmed by the server, as indicated by the confirmed field of the corresponding Channel.

pub async fn channel_open_direct_tcpip<A: Into<String>, B: Into<String>>( &self, host_to_connect: A, port_to_connect: u32, originator_address: B, originator_port: u32, ) -> Result<Channel<Msg>, Error>

Open a TCP/IP forwarding channel. This is usually done when a connection comes to a locally forwarded TCP/IP port. See RFC4254. The TCP/IP packets can then be tunneled through the channel using .data().

pub async fn channel_open_forwarded_tcpip<A: Into<String>, B: Into<String>>( &self, connected_address: A, connected_port: u32, originator_address: B, originator_port: u32, ) -> Result<Channel<Msg>, Error>

pub async fn channel_open_x11<A: Into<String>>( &self, originator_address: A, originator_port: u32, ) -> Result<Channel<Msg>, Error>

pub async fn exit_signal_request( &self, id: ChannelId, signal_name: Sig, core_dumped: bool, error_message: String, lang_tag: String, ) -> Result<(), ()>

If the program was killed by a signal, send the details about the signal to the client.

pub async fn disconnect( &self, reason: Disconnect, description: String, language_tag: String, ) -> Result<(), Error>

Allows a server to disconnect a client session