russh::server
Trait Server

pub trait Server {
    type Handler: Handler + Send + 'static;

    // Required method
    fn new_client(&mut self, peer_addr: Option<SocketAddr>) -> Self::Handler;

    // Provided methods
    fn handle_session_error(
        &mut self,
        _error: <Self::Handler as Handler>::Error,
    ) { ... }
    fn run_on_socket<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        config: Arc<Config>,
        socket: &'life1 TcpListener,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn run_on_address<'life0, 'async_trait, A>(
        &'life0 mut self,
        config: Arc<Config>,
        addrs: A,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where A: 'async_trait + ToSocketAddrs + Send,
             Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
}

Trait used to create new handlers when clients connect.

Required Associated Types

type Handler: Handler + Send + 'static

The type of handlers.

Required Methods

fn new_client(&mut self, peer_addr: Option<SocketAddr>) -> Self::Handler

Called when a new client connects.

Provided Methods

fn handle_session_error(&mut self, _error: <Self::Handler as Handler>::Error)

Called when an active connection fails.

fn run_on_socket<'life0, 'life1, 'async_trait>( &'life0 mut self, config: Arc<Config>, socket: &'life1 TcpListener, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Run a server on a specified tokio::net::TcpListener. Useful when dropping privileges immediately after socket binding, for example.

fn run_on_address<'life0, 'async_trait, A>( &'life0 mut self, config: Arc<Config>, addrs: A, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where A: 'async_trait + ToSocketAddrs + Send, Self: Send + 'async_trait, 'life0: 'async_trait,

Run a server. Create a new Connection from the server’s configuration, a stream and a Handler.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.