russh::client
Struct Session

pub struct Session { /* private fields */ }

Actual client session’s state.

It is in charge of multiplexing and keeping track of various channels that may get opened and closed during the lifetime of an SSH session and allows sending messages to the server.

Implementations

impl Session

pub fn channel_open_session(&mut self) -> Result<ChannelId, Error>

pub fn channel_open_x11( &mut self, originator_address: &str, originator_port: u32, ) -> Result<ChannelId, Error>

pub fn channel_open_direct_tcpip( &mut self, host_to_connect: &str, port_to_connect: u32, originator_address: &str, originator_port: u32, ) -> Result<ChannelId, Error>

pub fn channel_open_direct_streamlocal( &mut self, socket_path: &str, ) -> Result<ChannelId, Error>

pub fn request_pty( &mut self, channel: ChannelId, want_reply: bool, term: &str, col_width: u32, row_height: u32, pix_width: u32, pix_height: u32, terminal_modes: &[(Pty, u32)], )

pub fn request_x11( &mut self, channel: ChannelId, want_reply: bool, single_connection: bool, x11_authentication_protocol: &str, x11_authentication_cookie: &str, x11_screen_number: u32, )

pub fn set_env( &mut self, channel: ChannelId, want_reply: bool, variable_name: &str, variable_value: &str, )

pub fn request_shell(&mut self, want_reply: bool, channel: ChannelId)

pub fn exec(&mut self, channel: ChannelId, want_reply: bool, command: &[u8])

pub fn signal(&mut self, channel: ChannelId, signal: Sig)

pub fn request_subsystem( &mut self, want_reply: bool, channel: ChannelId, name: &str, )

pub fn window_change( &mut self, channel: ChannelId, col_width: u32, row_height: u32, pix_width: u32, pix_height: u32, )

pub fn tcpip_forward( &mut self, reply_channel: Option<Sender<Option<u32>>>, address: &str, port: u32, )

Requests a TCP/IP forwarding from the server

If reply_channel is not None, sets want_reply and returns the server’s response via the channel, Some<u32> for a success message with port, or None for failure

pub fn cancel_tcpip_forward( &mut self, reply_channel: Option<Sender<bool>>, address: &str, port: u32, )

Requests cancellation of TCP/IP forwarding from the server

If want_reply is true, returns a oneshot receiveing the server’s reply: true for a success message, or false for failure

pub fn send_keepalive(&mut self, want_reply: bool)

pub fn data(&mut self, channel: ChannelId, data: CryptoVec)

pub fn eof(&mut self, channel: ChannelId)

pub fn close(&mut self, channel: ChannelId)

pub fn extended_data(&mut self, channel: ChannelId, ext: u32, data: CryptoVec)

pub fn agent_forward(&mut self, channel: ChannelId, want_reply: bool)

pub fn disconnect( &mut self, reason: Disconnect, description: &str, language_tag: &str, )

pub fn has_pending_data(&self, channel: ChannelId) -> bool

pub fn sender_window_size(&self, channel: ChannelId) -> usize

pub fn remote_sshid(&self) -> &[u8]

Returns the SSH ID (Protocol Version + Software Version) the server sent when connecting

This should contain only ASCII characters for implementations conforming to RFC4253, Section 4.2:

Both the ‘protoversion’ and ‘softwareversion’ strings MUST consist of printable US-ASCII characters, with the exception of whitespace characters and the minus sign (-).

So it usually is fine to convert it to a String using String::from_utf8_lossy

impl Session

pub fn send_channel_msg(&self, channel: ChannelId, msg: ChannelMsg) -> bool

Send a ChannelMsg from the background handler to the client.