russh::client
Trait Handler

pub trait Handler: Sized + Send {
    type Error: From<Error> + Send + Debug;

    // Provided methods
    fn auth_banner<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        banner: &'life1 str,
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn check_server_key<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        server_public_key: &'life1 PublicKey,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_open_confirmation<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        id: ChannelId,
        max_packet_size: u32,
        window_size: u32,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_success<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_failure<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_close<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_eof<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_open_failure<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        reason: ChannelOpenFailure,
        description: &'life1 str,
        language: &'life2 str,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn server_channel_open_forwarded_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: Channel<Msg>,
        connected_address: &'life1 str,
        connected_port: u32,
        originator_address: &'life2 str,
        originator_port: u32,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn server_channel_open_agent_forward<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn server_channel_handle_unknown(
        &self,
        channel: ChannelId,
        channel_type: &[u8],
    ) -> bool { ... }
    fn server_channel_open_session<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn server_channel_open_direct_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        host_to_connect: &'life1 str,
        port_to_connect: u32,
        originator_address: &'life2 str,
        originator_port: u32,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn server_channel_open_x11<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: Channel<Msg>,
        originator_address: &'life1 str,
        originator_port: u32,
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn data<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        data: &'life1 [u8],
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn extended_data<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        ext: u32,
        data: &'life1 [u8],
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn xon_xoff<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        client_can_do: bool,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn exit_status<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        exit_status: u32,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn exit_signal<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        signal_name: Sig,
        core_dumped: bool,
        error_message: &'life1 str,
        lang_tag: &'life2 str,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn window_adjusted<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        new_size: u32,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn adjust_window(&mut self, channel: ChannelId, window: u32) -> u32 { ... }
    fn openssh_ext_host_keys_announced<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        keys: Vec<PublicKey>,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn disconnected<'life0, 'async_trait>(
        &'life0 mut self,
        reason: DisconnectReason<Self::Error>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }

}

A client handler. Note that messages can be received from the server at any time during a session.

Note: this is an async_trait. Click [source] on the right to see actual async function definitions.

Required Associated Types

type Error: From<Error> + Send + Debug

Provided Methods

fn auth_banner<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, banner: &'life1 str, session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when the server sends us an authentication banner. This is usually meant to be shown to the user, see RFC4252 for more details.

fn check_server_key<'life0, 'life1, 'async_trait>( &'life0 mut self, server_public_key: &'life1 PublicKey, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called to check the server’s public key. This is a very important step to help prevent man-in-the-middle attacks. The default implementation rejects all keys.

fn channel_open_confirmation<'life0, 'life1, 'async_trait>( &'life0 mut self, id: ChannelId, max_packet_size: u32, window_size: u32, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server confirmed our request to open a channel. A channel can only be written to after receiving this message (this library panics otherwise).

fn channel_success<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server signals success.

fn channel_failure<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server signals failure.

fn channel_close<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server closes a channel.

fn channel_eof<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server sends EOF to a channel.

fn channel_open_failure<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: ChannelId, reason: ChannelOpenFailure, description: &'life1 str, language: &'life2 str, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Called when the server rejected our request to open a channel.

fn server_channel_open_forwarded_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: Channel<Msg>, connected_address: &'life1 str, connected_port: u32, originator_address: &'life2 str, originator_port: u32, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Called when the server opens a channel for a new remote port forwarding connection

fn server_channel_open_agent_forward<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server opens an agent forwarding channel

fn server_channel_handle_unknown( &self, channel: ChannelId, channel_type: &[u8], ) -> bool

Called when the server gets an unknown channel. It may return true, if the channel of unknown type should be handled. If it returns false, the channel will not be created and an error will be sent to the server.

fn server_channel_open_session<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server opens a session channel.

fn server_channel_open_direct_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: ChannelId, host_to_connect: &'life1 str, port_to_connect: u32, originator_address: &'life2 str, originator_port: u32, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Called when the server opens a direct tcp/ip channel.

fn server_channel_open_x11<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: Channel<Msg>, originator_address: &'life1 str, originator_port: u32, session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when the server opens an X11 channel.

fn data<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: ChannelId, data: &'life1 [u8], session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when the server sends us data. The extended_code parameter is a stream identifier, None is usually the standard output, and Some(1) is the standard error. See RFC4254.

fn extended_data<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: ChannelId, ext: u32, data: &'life1 [u8], session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when the server sends us data. The extended_code parameter is a stream identifier, None is usually the standard output, and Some(1) is the standard error. See RFC4254.

fn xon_xoff<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, client_can_do: bool, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The server informs this client of whether the client may perform control-S/control-Q flow control. See RFC4254.

fn exit_status<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, exit_status: u32, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The remote process has exited, with the given exit status.

fn exit_signal<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: ChannelId, signal_name: Sig, core_dumped: bool, error_message: &'life1 str, lang_tag: &'life2 str, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

The remote process exited upon receiving a signal.

fn window_adjusted<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, new_size: u32, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the network window is adjusted, meaning that we can send more bytes. This is useful if this client wants to send huge amounts of data, for instance if we have called Session::data before, and it returned less than the full amount of data.

fn adjust_window(&mut self, channel: ChannelId, window: u32) -> u32

Called when this client adjusts the network window. Return the next target window and maximum packet size.

fn openssh_ext_host_keys_announced<'life0, 'life1, 'async_trait>( &'life0 mut self, keys: Vec<PublicKey>, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the server signals success.

fn disconnected<'life0, 'async_trait>( &'life0 mut self, reason: DisconnectReason<Self::Error>, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Called when the server sent a disconnect message

If reason is an Error, this function should re-return the error so the join can also evaluate it

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.