russh::server
Struct Session

pub struct Session { /* private fields */ }

A connected server session. This type is unique to a client.

Implementations

impl Session

pub fn handle(&self) -> Handle

Get a handle to this session.

pub fn writable_packet_size(&self, channel: &ChannelId) -> u32

pub fn window_size(&self, channel: &ChannelId) -> u32

pub fn max_packet_size(&self, channel: &ChannelId) -> u32

pub fn flush(&mut self) -> Result<(), Error>

Flush the session, i.e. encrypt the pending buffer.

pub fn flush_pending(&mut self, channel: ChannelId) -> usize

pub fn sender_window_size(&self, channel: ChannelId) -> usize

pub fn has_pending_data(&self, channel: ChannelId) -> bool

pub fn config(&self) -> &Config

Retrieves the configuration of this session.

pub fn disconnect( &mut self, reason: Disconnect, description: &str, language_tag: &str, )

Sends a disconnect message.

pub fn request_success(&mut self)

Send a “success” reply to a /global/ request (requests without a channel number, such as TCP/IP forwarding or cancelling). Always call this function if the request was successful (it checks whether the client expects an answer).

pub fn request_failure(&mut self)

Send a “failure” reply to a global request.

pub fn channel_success(&mut self, channel: ChannelId)

Send a “success” reply to a channel request. Always call this function if the request was successful (it checks whether the client expects an answer).

pub fn channel_failure(&mut self, channel: ChannelId)

Send a “failure” reply to a global request.

pub fn channel_open_failure( &mut self, channel: ChannelId, reason: ChannelOpenFailure, description: &str, language: &str, )

Send a “failure” reply to a request to open a channel open.

pub fn close(&mut self, channel: ChannelId)

Close a channel.

pub fn eof(&mut self, channel: ChannelId)

Send EOF to a channel

pub fn data(&mut self, channel: ChannelId, data: CryptoVec)

Send data to a channel. On session channels, extended can be used to encode standard error by passing Some(1), and stdout by passing None.

The number of bytes added to the “sending pipeline” (to be processed by the event loop) is returned.

pub fn extended_data( &mut self, channel: ChannelId, extended: u32, data: CryptoVec, )

Send data to a channel. On session channels, extended can be used to encode standard error by passing Some(1), and stdout by passing None.

The number of bytes added to the “sending pipeline” (to be processed by the event loop) is returned.

pub fn xon_xoff_request(&mut self, channel: ChannelId, client_can_do: bool)

Inform the client of whether they may perform control-S/control-Q flow control. See RFC4254.

pub fn keepalive_request(&mut self)

Ping the client to verify there is still connectivity.

pub fn exit_status_request(&mut self, channel: ChannelId, exit_status: u32)

Send the exit status of a program.

pub fn exit_signal_request( &mut self, channel: ChannelId, signal: Sig, core_dumped: bool, error_message: &str, language_tag: &str, )

If the program was killed by a signal, send the details about the signal to the client.

pub fn channel_open_session(&mut self) -> Result<ChannelId, Error>

Opens a new session channel on the client.

pub fn channel_open_direct_tcpip( &mut self, host_to_connect: &str, port_to_connect: u32, originator_address: &str, originator_port: u32, ) -> Result<ChannelId, Error>

Opens a direct TCP/IP channel on the client.

pub fn channel_open_forwarded_tcpip( &mut self, connected_address: &str, connected_port: u32, originator_address: &str, originator_port: u32, ) -> Result<ChannelId, Error>

Open a TCP/IP forwarding channel, when a connection comes to a local port for which forwarding has been requested. See RFC4254. The TCP/IP packets can then be tunneled through the channel using .data().

pub fn channel_open_x11( &mut self, originator_address: &str, originator_port: u32, ) -> Result<ChannelId, Error>

Open a new X11 channel, when a connection comes to a local port. See RFC4254. TCP/IP packets can then be tunneled through the channel using .data().

pub fn channel_open_agent(&mut self) -> Result<ChannelId, Error>

Opens a new agent channel on the client.

pub fn tcpip_forward( &mut self, address: &str, port: u32, reply_channel: Option<Sender<Option<u32>>>, )

Requests that the client forward connections to the given host and port. See RFC4254. The client will open forwarded_tcpip channels for each connection.

pub fn cancel_tcpip_forward( &mut self, address: &str, port: u32, reply_channel: Option<Sender<bool>>, )

Cancels a previously tcpip_forward request.

pub fn remote_sshid(&self) -> &[u8]

Returns the SSH ID (Protocol Version + Software Version) the client sent when connecting

This should contain only ASCII characters for implementations conforming to RFC4253, Section 4.2:

Both the ‘protoversion’ and ‘softwareversion’ strings MUST consist of printable US-ASCII characters, with the exception of whitespace characters and the minus sign (-).

So it usually is fine to convert it to a String using String::from_utf8_lossy