russh::server
Trait Handler

pub trait Handler: Sized {
    type Error: From<Error> + Send;

    // Provided methods
    fn auth_none<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        user: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn auth_password<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        user: &'life1 str,
        password: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn auth_publickey_offered<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        user: &'life1 str,
        public_key: &'life2 PublicKey,
    ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn auth_publickey<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        user: &'life1 str,
        public_key: &'life2 PublicKey,
    ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn auth_keyboard_interactive<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        user: &'life1 str,
        submethods: &'life2 str,
        response: Option<Response<'async_trait>>,
    ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn auth_succeeded<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_close<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_eof<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_open_session<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: Channel<Msg>,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn channel_open_x11<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: Channel<Msg>,
        originator_address: &'life1 str,
        originator_port: u32,
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn channel_open_direct_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: Channel<Msg>,
        host_to_connect: &'life1 str,
        port_to_connect: u32,
        originator_address: &'life2 str,
        originator_port: u32,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn channel_open_forwarded_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: Channel<Msg>,
        host_to_connect: &'life1 str,
        port_to_connect: u32,
        originator_address: &'life2 str,
        originator_port: u32,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn channel_open_confirmation<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        id: ChannelId,
        max_packet_size: u32,
        window_size: u32,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn data<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        data: &'life1 [u8],
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn extended_data<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        code: u32,
        data: &'life1 [u8],
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn window_adjusted<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        new_size: u32,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn adjust_window(&mut self, channel: ChannelId, current: u32) -> u32 { ... }
    fn pty_request<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        term: &'life1 str,
        col_width: u32,
        row_height: u32,
        pix_width: u32,
        pix_height: u32,
        modes: &'life2 [(Pty, u32)],
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn x11_request<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        single_connection: bool,
        x11_auth_protocol: &'life1 str,
        x11_auth_cookie: &'life2 str,
        x11_screen_number: u32,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn env_request<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        variable_name: &'life1 str,
        variable_value: &'life2 str,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn shell_request<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn exec_request<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        data: &'life1 [u8],
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn subsystem_request<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        name: &'life1 str,
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn window_change_request<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        col_width: u32,
        row_height: u32,
        pix_width: u32,
        pix_height: u32,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn agent_request<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn signal<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        channel: ChannelId,
        signal: Sig,
        session: &'life1 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn tcpip_forward<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self,
        address: &'life1 str,
        port: &'life2 mut u32,
        session: &'life3 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait { ... }
    fn cancel_tcpip_forward<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        address: &'life1 str,
        port: u32,
        session: &'life2 mut Session,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }

}

Server handler. Each client will have their own handler.

Note: this is an async_trait. Click [source] on the right to see actual async function definitions.

Required Associated Types

type Error: From<Error> + Send

Provided Methods

fn auth_none<'life0, 'life1, 'async_trait>( &'life0 mut self, user: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Check authentication using the none method. Russh makes sure rejection happens in time config.auth_rejection_time, except if this method takes more than that.

fn auth_password<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, user: &'life1 str, password: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Check authentication using the password method. Russh makes sure rejection happens in time config.auth_rejection_time, except if this method takes more than that.

fn auth_publickey_offered<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, user: &'life1 str, public_key: &'life2 PublicKey, ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Check authentication using the publickey method. This method should just check whether the public key matches the authorized ones. Russh then checks the signature. If the key is unknown, or the signature is invalid, Russh guarantees that rejection happens in constant time config.auth_rejection_time, except if this method takes more time than that.

fn auth_publickey<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, user: &'life1 str, public_key: &'life2 PublicKey, ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Check authentication using the publickey method. This method is called after the signature has been verified and key ownership has been confirmed. Russh guarantees that rejection happens in constant time config.auth_rejection_time, except if this method takes more time than that.

fn auth_keyboard_interactive<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, user: &'life1 str, submethods: &'life2 str, response: Option<Response<'async_trait>>, ) -> Pin<Box<dyn Future<Output = Result<Auth, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Check authentication using the keyboard-interactive method. Russh makes sure rejection happens in time config.auth_rejection_time, except if this method takes more than that.

fn auth_succeeded<'life0, 'life1, 'async_trait>( &'life0 mut self, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when authentication succeeds for a session.

fn channel_close<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the client closes a channel.

fn channel_eof<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the client sends EOF to a channel.

fn channel_open_session<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: Channel<Msg>, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a new session channel is created. Return value indicates whether the channel request should be granted.

fn channel_open_x11<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: Channel<Msg>, originator_address: &'life1 str, originator_port: u32, session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when a new X11 channel is created. Return value indicates whether the channel request should be granted.

fn channel_open_direct_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: Channel<Msg>, host_to_connect: &'life1 str, port_to_connect: u32, originator_address: &'life2 str, originator_port: u32, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Called when a new TCP/IP is created. Return value indicates whether the channel request should be granted.

fn channel_open_forwarded_tcpip<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: Channel<Msg>, host_to_connect: &'life1 str, port_to_connect: u32, originator_address: &'life2 str, originator_port: u32, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Called when a new forwarded connection comes in. https://www.rfc-editor.org/rfc/rfc4254#section-7

fn channel_open_confirmation<'life0, 'life1, 'async_trait>( &'life0 mut self, id: ChannelId, max_packet_size: u32, window_size: u32, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the client confirmed our request to open a channel. A channel can only be written to after receiving this message (this library panics otherwise).

fn data<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: ChannelId, data: &'life1 [u8], session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when a data packet is received. A response can be written to the response argument.

fn extended_data<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: ChannelId, code: u32, data: &'life1 [u8], session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Called when an extended data packet is received. Code 1 means that this packet comes from stderr, other codes are not defined (see RFC4254).

fn window_adjusted<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, new_size: u32, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when the network window is adjusted, meaning that we can send more bytes.

fn adjust_window(&mut self, channel: ChannelId, current: u32) -> u32

Called when this server adjusts the network window. Return the next target window.

fn pty_request<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: ChannelId, term: &'life1 str, col_width: u32, row_height: u32, pix_width: u32, pix_height: u32, modes: &'life2 [(Pty, u32)], session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

The client requests a pseudo-terminal with the given specifications.

fn x11_request<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: ChannelId, single_connection: bool, x11_auth_protocol: &'life1 str, x11_auth_cookie: &'life2 str, x11_screen_number: u32, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

The client requests an X11 connection.

fn env_request<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, channel: ChannelId, variable_name: &'life1 str, variable_value: &'life2 str, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

The client wants to set the given environment variable. Check these carefully, as it is dangerous to allow any variable environment to be set.

fn shell_request<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The client requests a shell.

fn exec_request<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: ChannelId, data: &'life1 [u8], session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

The client sends a command to execute, to be passed to a shell. Make sure to check the command before doing so.

fn subsystem_request<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, channel: ChannelId, name: &'life1 str, session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

The client asks to start the subsystem with the given name (such as sftp).

fn window_change_request<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, col_width: u32, row_height: u32, pix_width: u32, pix_height: u32, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The clients pseudo-terminal window size has changed.

fn agent_request<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The client requests OpenSSH agent forwarding

fn signal<'life0, 'life1, 'async_trait>( &'life0 mut self, channel: ChannelId, signal: Sig, session: &'life1 mut Session, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The client is sending a signal (usually to pass to the currently running process).

fn tcpip_forward<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 mut self, address: &'life1 str, port: &'life2 mut u32, session: &'life3 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Used for reverse-forwarding ports, see RFC4254. If port is 0, you should set it to the allocated port number.

fn cancel_tcpip_forward<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, address: &'life1 str, port: u32, session: &'life2 mut Session, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Used to stop the reverse-forwarding of a port, see RFC4254.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.